--- app/models/User.scala
@@ -1,9 +1,0 @@
-package models
-
-case class User(
-    name: String, 
-    last: String, 
-    permission: Option[Permission] = None) {
-
-  def nick = s"${name.toLowerCase}_${last.toLowerCase}"
-}
