+++ app/models/org/hablapps/codemotion14/user.scala
@@ -1,8 +1,16 @@
-package models
+package models.org.hablapps.codemotion14
 
 import play.api.cache.Cache
 import play.api.Play.current
 import play.api.Logger
+
+case class User(
+    name: String, 
+    last: String, 
+    permission: Option[Permission] = None) {
+
+  def nick = s"${name.toLowerCase}_${last.toLowerCase}"
+}
 
 object Users {
